Fix killing of location process

+++ transmission-cli/.transmission-cli/kill.sh
@@ -1,6 +1,6 @@
 #!/bin/sh
 killall transmission-cli
-location_pid=pgrep -f "cat /dev/location"
-kill $location_pid
-
-
+location_pid=$(pgrep -f "cat /dev/location")
+if [ -n "$location_pid" ]; then
+  kill $location_pid
+fi
